+++ src/booklab/booklabd/__init__.py
@@ -20,14 +20,14 @@ from booklab import CONF_ROOT
 from booklab import DATA_ROOT
 
 
-template_location = os.path.abspath("../../../docs")
+template_location = os.path.abspath("../../../docs/")
 # create web application object
 api_app = init_app(
     __name__,
     static_site_dir = template_location
 )
 
-# 4dbg... print(f"*** booklabd.__init__ Created {api_app=} with {template_location=}")
+print(f"*** booklabd.__init__ Created {api_app=} with {template_location=}")
 
 # create database object by opening JSON files
 db_books, db_system = init_db()