+++ src/booklab/booklabd/routes.py
@@ -7,11 +7,10 @@ Author: Petre Iordanescu (petre.iordanescu@gmail.com)
 Created: 250825
 """
 
+from flask import render_template
 from booklab.booklabd import api_app
 from booklab.booklabd import db_books, db_system
 
-
-# 4dbg... print(f"*** booklabd.routes.py *** here is api_app variable {api_app}")
 
 @api_app.route("/api/bcat/")
 def api_bcat():
@@ -26,12 +25,16 @@ def api_bcat():
     bcat_records = db_books.getAll()
     if not (type(bcat_records) == type(list())):
         bcat_records = list().append(bcat_records)  # make it list if is not (when just 1 record)
-    str_bcat_records = f"{bcat_records=}"  # 4dbg...
+    #...4dbg str_bcat_records = f"{bcat_records=}"  # 4dbg...
 
     #... #TODO render template as nxt line
     # Fkask_render_template_function_w.params (bcat_data=bcat_records)
 
-    ret_str = "bcat records is:<br>" + str_bcat_records  # 4dbg
+    #..;4dbg ret_str = "bcat records is:<br>" + str_bcat_records  # 4dbg
+    ret_str = render_template(
+        "bcat/bcat.html",
+        bcat_data=bcat_records
+    )
     return ret_str